add handleNewCommit
