+++ src/components/Comment.js
@@ -1,29 +1,93 @@
 import React from "react";
-// import Moment from "react-moment";
-import { Link } from "@reach/router";
-import DefaultProfileImg from "../images/profile.jpg";
+import cx from "clsx";
+import { makeStyles } from "@material-ui/styles";
+import Box from "@material-ui/core/Box";
+import Link from "@material-ui/core/Link";
+import Card from "@material-ui/core/Card";
+import CardContent from "@material-ui/core/CardContent";
+import CardMedia from "@material-ui/core/CardMedia";
+import Divider from "@material-ui/core/Divider";
+import Rating from "@material-ui/lab/Rating";
+import ArrowForwardIos from "@material-ui/icons/ArrowForwardIos";
+import Favorite from "@material-ui/icons/Favorite";
+import { usePushingGutterStyles } from "@mui-treasury/styles/gutter/pushing";
+import { useLabelIconStyles } from "@mui-treasury/styles/icon/label";
+import { useRowFlexStyles } from "@mui-treasury/styles/flex/row";
 
-const Comment = ({ date, profileImage, message, username, messageId }) => (
-  <li className="row mb-4 message-area py-2">
-    <div className="col-lg-2 d-flex align-items-center">
-      <img
-        src={profileImage || DefaultProfileImg}
-        alt={username}
-        height="100"
-        width="100"
-        className="timeline-image d-none d-lg-block"
-      />
-    </div>
-    <div className="col-12 col-lg-10">
-      <div className="w-100 d-flex mb-1">
-        <Link className="username" to="/">
-          @{username}
-        </Link>
-        {/* <span className="text-muted ml-auto">{date}</span> */}
-      </div>
-      <p className="message-body">{message}</p>
-    </div>
-  </li>
-);
+const useStyles = makeStyles(({ spacing, palette }) => ({
+  card: {
+    display: "flex",
+    padding: spacing(2),
+    borderRadius: 16
+  },
+  media: {
+    minWidth: "25%",
+    maxWidth: "25%",
+    flexShrink: 0,
+    backgroundColor: palette.grey[200],
+    borderRadius: 12,
+    marginRight: "2rem"
+    // boxShadow: "0 2px 8px 0 #c1c9d7, 0 -2px 8px 0 #cce1e9"
+  },
+  rating: {
+    verticalAlign: "text-top"
+  },
+  content: {
+    padding: spacing(0, 2, 0, 0)
+  },
+  heading: {
+    fontSize: 17,
+    fontWeight: "bold",
+    letterSpacing: "0.5px",
+    marginBottom: 0,
+    marginRight: spacing(1.5),
+    display: "inline-block"
+  },
+  body: {
+    fontSize: 14,
+    color: palette.grey[500]
+  },
+  divider: {
+    margin: spacing(1, 0)
+  },
+  textFooter: {
+    fontSize: 14
+  },
+  icon: {
+    fontSize: "1.2rem",
+    verticalAlign: "bottom"
+  }
+}));
+
+const Comment = ({ username, message, profileImage }) => {
+  const styles = useStyles();
+  const gutterStyles = usePushingGutterStyles({ space: 1.5 });
+  const labelStyles = useLabelIconStyles({ linked: true });
+  const flexStyles = useRowFlexStyles();
+  return (
+    <Card className={styles.card} variant="outlined">
+      <CardMedia className={styles.media} image={profileImage} />
+      <CardContent className={styles.content}>
+        <Box mb={1}>
+          <h3 className={styles.heading}>{username}</h3>
+          <Rating name={"rating"} value={2} className={styles.rating} size={"small"} />
+        </Box>
+        <p className={styles.body}>{message}</p>
+        <Divider className={styles.divider} light />
+        <div className={flexStyles.parent}>
+          {/* eslint-disable-next-line jsx-a11y/anchor-is-valid */}
+          <Link className={cx(labelStyles.primaryLink, styles.textFooter)} component={"button"}>
+            Read more <ArrowForwardIos className={labelStyles.icon} />
+          </Link>
+          <div className={cx(flexStyles.rightChild, flexStyles.parent, gutterStyles.parent)}>
+            <button type={"button"} className={labelStyles.link}>
+              <Favorite className={labelStyles.icon} /> 12
+            </button>
+          </div>
+        </div>
+      </CardContent>
+    </Card>
+  );
+};
 
 export default Comment;
